not prllx switch header section
Not really parallax work but replacing nav with the title section first and ensuring animations play nicely. Devised a new theme that matches grid areas and could be extended/renamed to be useful throughout the site. Definitely the code needs some tidying/refactoring and storybook needs adding for easier development.

The next stage is getting the page component to match the colour of the grid area.